proc: add parent/child hierarchy locking
The locking order is always parent -> child.  Since the root process can never
be another process's child, when it is involved the order is root -> parent ->
child.  This requires a new helper that binds and locks a process's parent with
a retry loop in case of a concurrent re-parent.

Most accesses naturally only touch the current process, or its children, and
therefore don't need this.  When signalling the parent (e.g. due to the process
stopping or continuing), we have to do this relock/binding dance.  This creates
a bit of awkwardness in those parts of the code.

I'm overall pretty unhappy with the state of proc_prep_user_return().  It is
logically racy w.r.t. signal handling already, and this new locking just makes
it even more awkward.