proc: handle SIGCONT in dispatch_signal() with SIGSTOP
This consolidates stop and continue handling in one place.  The asymmetry
originally was due to the following:
 1) SIGSTOP is _not_ the only stopping signal, and the others _can_ be
    masked/handled.
 2) SIGSTOP cannot be masked or handled
 3) SIGCONT _can_ be masked/handled, but should still continue the process
 4) SIGCONT is the only continuing signal

(1) means that stopping is naturally handled in dispatch_signal(), since
multiple signals can trigger it.  (2) means that SIGSTOP can be handled there as
well.  Meanwhile, (3) means that handling SIGCONT in dispatch_signal() via the
default action is difficult, as it must continue the process even if masked or
handled.  And (4) means we can just special-case SIGCONT outside the
dispatch_signal() loop (as before).

Now, we create a new internal signal, SIGAPOS_FORCE_CONT, which is similar to
SIGSTOP in that it can't be masked, ignored, or handled --- if generated, it
will always take its default action (to continue the process).  Whenever we
generate SIGCONT, we also generate SIGAPOS_FORCE_CONT.  SIGCONT will be handled
however it needs to be (which could be masked or dispatched to a handler), but
SIGAPOS_FORCE_CONT will force the process to continue then anyway.

This causes an unfortunate side effect: processes that have a thread blocking in
a syscall will have that syscall interrupted to handle the SIGCONT.  Previously
this wouldn't happen if SIGCONT was masked, but now SIGAPOS_FORCE_CONT will
cause the syscall to be interrupted even if SIGCONT is masked.  Neither the old
nor new behavior is optimal (ideally the process is continued without
interrupting still-blocking syscalls), but it's now more visible than before.
Tests added for this scenario.

The proc_prep_user_return() loop is pretty awkward now, but will be rewritten
soon.