proc: fix signal assignment race condition in user return
This lifts the lock/unlock out of the signal handling code, allowing us to
atomically assign/check signals and dispatch them (at the small cost of doing
the context extraction with the process lock held).  This is not functionally
required, but is a bit cleaner.

The second, more important, change is pull dispatch action handling out of
proc_dispatch_pending_signals() so that the caller can retry assigning signals
before sleeping.  The STOPPED sleep will only happen once we've finished a cycle
of dispatching signals that didn't require us to unlock the process's spinlock.