memory: pre-allocate 5 pages of the heap at kmalloc init
The next change will introduce locking to vm_handle_page_fault().  When a page
fault happens before process initialization is complete, the kernel
double-faults when it tries to take a lock.  Early in kernel boot, we don't
actually need to lock as there will never be multiple threads or SMP.

Rather than making lock-taking conditional on boot stage (which adds overhead to
all lock sites, and would be complex as it would require handling anywhere in
the VM subsystem that takes the process mutex), this change makes kmalloc
preallocate the first N pages of the heap.  This allows that much memory to be
allocated and used without incurring any early-boot page faults.