memory: pre-map all allocated heap pages
Rather than returning an address then letting the caller fault when it accesses
it, this ensures that all memory returned by kmalloc() is backed by physical
pages, and there won't be any page faults.  This allows locking in the page
fault handler (and is just more efficient).  It also lets us completely get rid
of the non-memobj-backed branch in the page fault handler.

If the calls to preallocate_block() are removed, the kernel can't even boot ---
but add a basic kmalloc test that runs all the time anyway.