proc: protect vm_area_list with process->mu
This is mostly a mechanical change with the preceding kmalloc changes already
made.  This adds a check that VM global kernel mappings are only created before
the first fork() --- after that, they won't be propagate correctly.  This allows
us to set them up without the process lock held, which is required as it happens
early during boot.