proc: fix locking in exec() and document user_arch locking
user_arch is only ever modified in do_execve(), at which point it is guaranteed
that no other process threads are running.  Therefore, all in-process uses of it
(e.g. in syscalls) can read without a lock as it won't be changing.  Only
out-of-process reads must lock to avoid concurrent modification with do_execve()
--- the only current such reader is procfs.

This also fixes locking in do_execve(), and updates process->execed under lock.