tsan: add test for over-synchronizing on acquire-load
There is an existing bug in the TSAN atomic code in that the load and acquire
operations (and store+release) are not done atomically.  This will be fixed
soon.  A naive fix is to simply reverse the load/acquire order (so that rather
than acquire first then load, we load first then acquire) --- this introduces an
opposite bug that generates false negatives rather than false positives, if we
end up acquiring a clock that's "later" than the load we actually just made.

This new test attempts to validate that scenario, but does so in a somewhat
brittle way.  It is currently non-flaky, but could be flaky down the road with
SMP as the race will no longer deterministically occur.