tsan: fix acq/rel load/store race
Currently, the acquire/release operations on the sync object are not done
atomically with the actual load/store/RMW.  This introduces false positives and
negatives (depending on the ordering).

Looking at load, the race can cause false positives if the thread is preempted
in between the check/acquire and the actual load, which trigger rarely with the
existing TSAN atomic tests.  The sequence of events is:
 1) T1 checks atomic access, does acquire on the flag (sync object)
 2) T2 writes non-atomic data
 3) T2 does atomic release write of 1 to the flag (sync object)
 4) T1 reads the flag, getting the value written in (3)
 5) T1 sees the flag is 1, then reads non-atomic data --> false positive

The first thread sees the new value of the flag (1) but hasn't acquired the
clock values associated with the write of that value.  This can be seen (with
additional logging) in sponge-20250310-215201-riscv64-clang-0001.

The fix adds a spinlock to each sync object, and does the actual operation with
the lock held (for non-relaxed accesses).  The specific structure of these
functions (in particular the ordering of the relaxed fast path, and the
double-read on load) is modeled after clang's implementation, which I assume has
been optimized for performance.

The atomic load/store implementations can be turned into macros similar to
DEFINE_ATOMIC_RMW() down the road if we need to support more atomic types
besides uint32_t.