proc: disable TSAN for core defint/thread/proc state changes
Each of these functions changes the current thread execution state (for example,
which thread is running, or whether we're running a defint), which confuses TSAN
when it instruments accesses that happen inside the functions.

Disabling TSAN for the whole functions is unfortunately a bit course-grained,
but there's no obvious way to disable it only for a few lines of code without
factoring those out into a separate (smaller) function.  That might be the right
thing to do eventually.

This is (finally!) the last change required before supporting full-kernel TSAN.
With this patch, all kernel tests (except the tsan tests) can be run under
full-kernel TSAN with no races detected, though they don't all pass (due to
timing issues).