vfs: use atomic for vnode refcount
This is only a partial change to lock-free refcounts.  We still use the
g_vnode_cache_lock spinlock to protect decrements, and therefore the complex
logic in vfs_put().  The only functional change here is that we can new
increment the refcount in vfs_get() without taking the global lock.

All accesses are still relaxed since they only must be done atomically, and
don't (on their own) signal lifetime changes or synchronize other accesses.

This could potentially be replaced fully with refcount_t, but that would require
more logic to handle vnodes being in the table with a refcount of zero (or count
the reference in the table as a refcount itself).