riscv: define interrupt functions as inlineable
When running under TSAN_FULL, this shaves test time down by 15-25% (as measured
on a few runs of the vfs_mount and tcp tests).  Such a large amount of time is
spent enabling and disabling interrupts in the TSAN code that the savings are
substantial.  Because the TSAN code doesn't do anything else when
enabling/disabling interrupts (unlike normal code, which must call TSAN),
inlining is a big win.

This could be done for x86 targets as well (and probably should be), though I
expect the relative gains to be much smaller when not running under TSAN.