proc: avoid queuing multiple defint timer defints
This generally shouldn't be an issue, but when running under TSAN things can
pile up enough that we get into a death spiral and eventually exceed the max
defint queue length.

There's probably a clever way to do this using only g_defint_timer_next and not
a separate flag, but the correct interaction with defint_timer_create() isn't
immediately obvious to me.