tcp: handle fully-subsumed OOO queued packets correctly
Currently if a packet in the OOO queue is fully subsumed by an incoming packet
(the new packet covers the entire queued packet), the kernel will panic in
tcp_handle_data() at
  KASSERT_DBG(md->data_len >= trim_start);

Since the trim length will be longer than the packet itself.

This commit adds a re-validation of each queued OOO packet before it is
dispatched; if it is (now) invalid, it is simply dropped.