net: add exponential backoff to loopback packet handling
The static 100ms delay is insufficient when running the multi-threaded TCP tests
under TSAN.  They will occasionally get stuck with a backup of packets, usually
when two sockets are talking and have each dropped a packet in one direction,
and one has a zero receive window.  They get into a packet spraying fight (which
should be fixed separately), and because everything is running slower due to
TSAN, this starves the threads that would be reading data and opening up the
receive window (which would allow the zero-window socket to accept the other
socket's retransmit, which would ACK the missing data from the first socket,
unblocking the connection).

There are other pathologies (besides a zero-window starvation problem) that
cause the packet war when running under TSAN, but adding this dynamic backoff
seems to make them all go away.