kthread: use refcount for join/detach
Having a proper thread refcount means we can get rid of a lot of the special
join/detach logic.  Detaching a thread is now just a wrapper around
kthread_unref(), and it's safe for anyone who holds a reference to the thread to
join it.  Threads will now usually be cleaned up from the reap queue rather than
a joiner, even if they're not detached.

In practice, this only makes a difference for tests, because kernel threads are
never joined outside tests.  This was never really the right API for kernel
threads, in retrospect.