tsan: support lock init from interrupt context
This would be very strange for a normal lock (spinlock, mutex, etc) but can
happen for dynamic locks in sync objects if the sync object is first created
from an interrupt context.  Concretely, this is likely to happen on a thread's
refcount decrement.

I'm not sure why I originally restricted this to only work in thread or defint
contexts, perhaps just as a sanity check --- the logic is perfectly correct to
run in an interrupt context.