tcp: heap-allocate test state in large test functions
These functions are on the border of the 5000-byte limit on riscv64, and a
change that adds a small spinlock to kthread_queue_t is enough to push them
over.  Since these are individual test cases, they can't easily be split --- but
the test state structs (almost 1kbyte each) can be heap-allocated.