tsan: map metadata region for thread0 stack
The proper approach would be to do this for all .init-* data sections (data,
rodata, bss, etc), but initial_kstack is currently the only symbol we need to do
this for.  Currently no threads ever do a TSAN access on the thread0 stack, but
another change I have coming will cause them to (which, without this change,
causes a crash as there is no metadata region mapped).