kthread: make runnable a (relaxed) atomic
We want to read the runnable field in inner loops when popping threads from
queues.  Having it be protected by a future kthread_t spinlock would complicate
that locking significantly.  Instead, make this an atomic and access it only
with relaxed operations --- the enable/disable bit is only used in tests to
enforce particular event sequences, and does not require any ordering
guarantees.