usb: run HCD callbacks in a defint context
This is a very minimal change that allows us to make kthread pools (which are
only used by USB) not-interrupt-safe.  The UHCI code should really be updated to
do all the hard work from the interrupt handler in the tasklet instead, but
that's a deeper change --- and the USB code overall is a bit of a mess I don't
want to update right now.

I wrote this with Gemini, which noticed that there is a memory leak here (we
never free the allocated TDs, QHs, etc).  Cool!  Added a TODO.