proc: fix SMP race with cancelled defint timers
Once SMP exists, this code would be racy --- a thread could cancel a timer and
reuse the defint_timer_t with another call to defint_timer_create() while
another thread is about to run the callback (which relies on the existing
defint_timer_t struct for the temporary linked list).

Now, rather than reusing the defint_timer_t between when the timer is taken off
the global list and when its callback is run, we copy the callback to a
stack-allocated array in batches.

Users still must ensure it is safe for the callback to be run even after they
call defint_timer_cancel(), but now they are able to safely manage the lifetime
of the defint_timer_t memory (and in particular reuse it for another timer).