proc: support out-of-order spinlock unlocks
The actual use case for this will be doing special cross-thread spinlock unlocks
during context switching, when I want to lock two threads before the context
switch, then unlock them (on the same logical core) in the new thread _after_
the context switch.

This requires saving the lock state explicitly in the stack frame of the threads
so that the correct per-core state is restored later when the thread that is
yielding is rescheduled (and it will, in fact, likely be unlocking a different
lock than the one that it locked before it yielded.