common: add basic per-cpu data mechanism
This is a simple abstraction over declared variables, inspired by the way Linux
does per-CPU data.  Currently, as SMP is not implemented, the CPU index is
always 0.

For SMP platforms, per_cpu_init() will later have to allocate memory for each
actual CPU (aligned appropriately) and store the offsets in the array.