dev: add qemu profiler driver, and profile tests
This is a driver for the synthetic customer qemu profiler I've added to qemu.
This enables profiling automatically for all test runs (when a profiler device
is found).

Note: with gcc, increases kernel-measured elapsed time from 35s to 42s, but
actual host measured elapsed time stays mostly the same (45-46s)