arch: define arch_cpu_index() inline
This always returns 0 currently.  According to a qemu-based profile, we spend
about 4% of test time inside this function.  Doing a 20-round before/after
comparison running the TCP tests on riscv64/clang, this change shaves 110ms off
(~2.7%), but I think the error bars on that are pretty broad.

Strangely, in previous profiles arch_cpu_index() was showing up as _25%_ of TCP
tests.  Now that I am testing this change specifically, I can't reproduce that.