proc: add thread locking to kthread_switch()
This requires storing the "actual" last thread in a thread-local variable, so
that we can unlock it after the switch.  The old thread stored in the stack
frame will be the thread that the current thread previously switched _to_, which
isn't the same.

Disabling thread safety analysis for these functions, as modeling the behavior
across a thread switch is difficult to do without strange contortions.

I experimented with an optimization so that we don't disable interrupts in the
second spinlock locking, and don't try to enable interrupts in the first
spinlock unlocking.  Despite the fact that (according to profiling) we spend a
significant amount of time in those enabling/disabling interrupts, the
optimization didn't seem to have a material effect on overall runtime.