proc: merge spinlock implementations
This is not a functional change, but means that the two types can be safely cast
to each other.  This will allow some shenanigans I want to do in kthread to make
accessing the join_list thread-safe.  A careful user can cast between them (for
example, to call scheduler_wait_on on an interrupt-safe spinlock).