proc: protect thread state with spinlock
What we've been working towards.

This is mostly straightforward, but requires some shenanigans with join/exit,
which are currently both logically- and data-racy.  Now join_list is properly
protected with the spinlock, and kthread_join() atomically checks the target
thread's state and waits until it has exited.

In an SMP context it's possible for a thread that is currently running to be
selected to run on another core before it yields; that will need to be handled.