proc: set state to KTHREAD_PENDING in kthread_switch()
This means that a thread will remain in KTHREAD_RUNNING until right before it
actually yields.  In SMP, another thread could observe the thread in
KTHREAD_PENDING after it has enqueued itself to yield but before it actually
does.  Now, a thread cannot observe KTHREAD_PENDING while the thread itself is
running, because the state is only changed after we lock the spinlock (and don't
unlock it until after the yield is complete).

In non-SMP, this doesn't matter.  Even with SMP, it _usually_ doesn't matter,
except for a condition like the following:
  1) thread A on core 1 yields, waiting on queue X
  2) thread B on core 2 (simultaneously) wakes up queue X, moving thread A to
     the run queue
  3) thread B (or another core) pulls thread A off the run queue and tries to
     run it, even though it is still running on core 1.

With this change, in (3) the other core will see that thread A is still in
KTHREAD_RUNNING, and schedule the idle thread instead, which should be more than
enough time for core 1 to finish switching away from thread A.

This should be very rare in practice.