proc: use thread lock in scheduler_wait()
This unveiled a couple of hidden race conditions that were masked because
interrupts used to be disabled through the whole function.  In particular, a
race that would happen with SMP if the timeout interrupt/defint fired
simultaneously with the thread yielding.  That case is now handled explicitly.

There is still an argument to be made that interrupts should simply be disabled
the whole way through for simplicity, but this makes it clearer that the
semantics apply to SMP as well.

The timeout code currently uses thread->interruptable to determine if it should
actually cancel the wait; this may be redundant now with the check below.  I've
added a TODO to have a hard think on that and remove it if so.