boot: fix per-cpu initialization race
This occurs if an interrupt fires after interrupts are enabled but before
per_cpu_init() runs.  The interrupt handler accesses PER_CPU(g_current_thread),
which ends up calculating (and dereferencing) the address of the per-CPU var as
0.  Therefore, the thread pointer will be whatever is in memory at address 0.

This doesn't manifest on riscv, because no interrupts fire until timers are
enabled.  On x86-64, however timer interrupts can (for some reason) fire earlier
than that.

It so happens that address 0 is accessible, and contains a random strange
address, which is interpreted as the thread pointer --- then the kernel either
faults accessing that address (likely), or accesses something random and
corrupts it (and fails a later assertion).

Fix this by moving per_cpu_init() to be before we actually initialize any data
structures.  We could also do this by keeping interrupts disabled longer, but
this is safer.

This can be easily repro'd by adding
  for (volatile int i = 0; i < 100000000; ++i);
after enable_interrupts() in kmain().

This bug goes back to when PER_CPU was written, but first manifests in 0003fc756
("kthread: make g_current_thread per-CPU with PER_CPU()").