proc: use spinlock/per-cpu for defint state
This replaces defint state with spinlock-protected global state and per-cpu
variables.  I used an atomic for g_defints_enabled because it is accessed by
interrupts --- I'm not positive this is strictly necessary, we're already
outside normal defined behavior anyway.  For g_defint_running, it's only
accessed with the lock held anyway, so I don't bother with an atomic.