proc: remove bad kthread->queue == NULL assertion
This was originally added as insurance to make sure we were keeping the thread
metadata up to date --- a sanity check that the current thread's queue was NULL.
There is nothing specific about scheduler_interrupt_thread() that requires that
the current thread isn't on a queue.

This assertion is now incorrect.  If scheduler_interrupt_thread() is called from
a defint context while the current thread is in the process of yielding, it can
observe thread->queue to be non-NULL.  See for example
sponge-20250811-215610-riscv64-clang-0001:
Thread 4 (crashing):
 #0 0xffffffff8020ab28 in die() [kassert.c:58]
 #1 0xffffffff8020ac5a in kassert_msg() [kassert.c:73]
 #2 0xffffffff80234380 in scheduler_interrupt_thread() [scheduler.c:108]
 #3 0xffffffff80238d1e in do_assign_signal() [signal.c:175]
 #4 0xffffffff802388ac in proc_try_assign_signal() [signal.c:197]
 #5 0xffffffff802385fa in proc_force_signal_locked() [signal.c:215]
 #6 0xffffffff8043ca26 in timer_cb() [alarm.c:34]
 #7 0xffffffff80230552 in defint_timer_defint() [defint_timer.c:86]
 #8 0xffffffff80230268 in defint_process_queued() [defint.c:116]
 #9 0xffffffff80230110 in defint_set_state() [defint.c:77]
 #10 0xffffffff80236d0e in kspin_unlock() [spinlock.c:95]
 #11 0xffffffff80234d4c in scheduler_wait() [scheduler.c:325]
 #12 0xffffffff80235220 in scheduler_wait_on_splocked() [scheduler.c:382]
 #13 0xffffffff804328dc in nvme_submit_blocking() [controller.c:344]
 #14 0xffffffff804ad654 in nvme_bd_op() [block_dev.c:78]
 #15 0xffffffff804ad37a in nvme_bd_write() [block_dev.c:98]
 #16 0xffffffff8043505a in bd_write_page() [memobj_block_dev.c:108]
 #17 0xffffffff802177fa in flush_cache_entry() [block_cache.c:181]
 #18 0xffffffff802158ca in flush_queue_thread() [block_cache.c:245]
 #19 0xffffffff80232e4e in kthread_trampoline() [kthread.c:134]
 #20 0xffffffff80423360 in kthread_arch_init_thread() [kthread.c:58]
PANIC: assertion failed: kthread_current_thread()->queue == 0x0 (proc/scheduler.c:108)

It also causes spurious TSAN failures, because the yielded-from thread writes
thread->queue (to add it to the run queue) but there is not a synchronization
event with the newly-running thread, even though it takes the thread lock.
There arguably should be, but that's a separate issue (and gets into messiness
around correct instrumentation of kthread/scheduler code preventing TSAN from
being useful for the rest of the kernel due to over-synchronization).  See
sponge-20250820-102038-riscv64-clang-0001 for an example of this.