proc: implement kmutex with raw spinlock
This makes kmutex_t SMP-safe, and removes another major usage of
PUSH_AND_DISABLE_INTERRUPTS().  This required a bunch of work --- in particular
to implement the per-thread spinlock to protect kthread_t::queue.

I originally tried this with a normal spinlock but that causes a major slowdown
(2x runtime for the VFS tests).  With the raw spinlock, it actually speeds the
VFS tests _up_ --- though that's because of getting rid of the dynamic interrupt
state cleanup verification in PUSH_AND_DISABLE_INTERRUPTS().

This also removes a major source of TSAN legacy-sync events in the scheduler, as
all kmutex_lock() calls now block on a lock rather than passing in NULL.