proc: fix legacy TSAN sync in scheduler_wait() for preemptible threads
This was broken recently when I added sched_disable_preemption() to
scheduler_wait(), instead of disabling interrupts.  The inner check for whether
the current thread is preemptible now would always view the thread as
NON-preemptible, which means that every single call to ksleep() or
scheduler_yield() would do a legacy full-sync, even if the thread is
preemptible.

This breaks all the TSAN tests (which rely on ksleep()/scheduler_yield() to
trigger race conditions), though they are still broken for other
over-synchronization reasons even with this fix.