proc: add no-defints-queued fast path to avoid false synchronization
Without this, every time a spinlock is unlocked, there is a global
sychronization when the defint queue is checked, causing false negatives.  This
adds a non-synchronizing fast path check first that avoids that synchronization
point in TSAN if there aren't any defints scheduled.

I needed to update the existing test to trigger the false negative by adding an
empty critical section.  I did the same for the mutex and interrupt-safe
spinlock tests as well for good measure.

Note that TSAN tests are still broken for other over-synchronization reasons.
Even with this, the "defint race (interrupt races with interrupt-ctx defint #2)"
test is still broken, as there is false sharing between the current interrupt
context and the defint context when the defint spinlock is unlocked right before
running the defint.

A better fix would be to opt defint code entirely out of TSAN synchronization,
as with other core code I'm working to fix.