proc: fix SMP race on thread exit
TSAN (with some uncommitted changes) flagged a false positive data race on the
thread's stack, in between where it is read during thread exit (right before the
switch) and where it is freed by the next thread.

This doesn't fix that, but looking at it made me realize there is a logical race
on thread exit --- the exiting thread adds itself to the reap queue, then
unlocks everything and switches.  If another thread switch happens on a
different CPU simultaneously, that other CPU could observe the first thread on
the reap queue and racily free it before it has actually exited.

The reap queue isn't actually even necessary at this point (though it dates back
to one of the earliest commits to the kernel!).  This change eliminates it, and
just has the next thread that is switched to check if the switched-from thread
has just exited, and cleans it up if so.

It is slightly awkward to have the duplicated code in two places but I don't see
a refactor that's worth the squeeze right now.