tsan: add TSAN_CORE option
This isn't functional yet, but will be a mode where TSAN is enabled for _all_
kernel code, including core scheduler/thread/interrupt code.  This means that
data races in those parts of the kernel can be found, at the cost of making it
far less likely to find races everywhere else.