tsan: don't instrument busy_loop() in tests
Not only is instrumenting the busy loops slow, but it also risks blowing out the
access logs of the thread, which makes one of the interrupt tests flaky when run
under TSAN_FULL (race is correctly caught, but the access isn't found in the
logs).

This removes the special-casing for TSAN_FULL altogether.