tsan: deflake a few tests under TSAN_FULL
There are several flakes caused by two things:
 * false sharing between adjacent fields (under TSAN_FULL)
 * full history of test memory causing replacement of important accesses

This solves both of them by clearing test memory's history before each test, and
aligning the test fields such that they won't share shadow cells.  Just the
history-clearing is sufficient (there is false sharing, but not enough accesses
to cause overwriting), but doing both for completeness.

These failures manifest in the multi-thread atomic value tests, and show up as
either a failure to find an expected race (if a false-sharing safe read replaces
the racy previous write before the race is triggered), or a failure to find the
access in the access history.

Fix a couple other small things while I'm in here.