dev: remove dynamic interrupt-state cleanup check
This is no longer really necessary, because,
1) almost all uses of PUSH_AND_DISABLE_INTERRUPTS() have been replace
2) thread safety static analysis will find mismatched instances at compile time

The dynamic check slows things down substantially under TSAN, and also causes
false positives on the thread stack (due to calls to
PUSH_AND_DISABLE_INTERRUPTS() etc from TSAN_CORE functions).  The "full" fix for
that would be to wipe the access history of the entire thread stack when the
thread is destroyed (as we [will] do with the thread object) --- but that denies
up the opportunity to catch races after thread exit, and this seems to be
sufficient.