tsan: add unit tests for tsc_* functions
This tests each of the tsc_* spinlock functions individually, which is a bunch
of tests but not much test runtime (~100ms).  Note that the tests don't pass
properly at HEAD, but do work with my pending changes that use tsc_* functions
in the scheduler core.

This also adds instrumentation for the access type to expect_report() to give
better failure messages.