proc: disable TSAN for scheduler code
This is mostly mechanical replacement of spinlock functions, but we also disable
TSAN dynamically during scheduler_wait().  This avoids over-synchronization when
we check the signal status of the process.

This is sufficient to un-break the tsan tests under TSAN_LIB!  They still don't
work under TSAN_FULL, though (that requires some more changes).