tools: add batch optimization to symbolize.py
Each call to addr2line costs ~20ms (very approximately).  When symbolizing a
live log, this isn't really a problem.  When processing an existing log file
(such as when symbolizing it for recording in sponge), it adds up.

This adds a batch mode to symbolize.py, where it gathers many lines up then
symbolizes all the addresses at once.  On a sample 2400-line log file with many
stack traces (from a run of the tsan tests), symbolizing goes from 4s to 0.2s.

I tried originally to make this work automatically by putting a timeout in with
select() (so when a certain amount of time passes without reading a line, we end
the chunk and write out everything we've found) --- however this doesn't work
well, because sys.stdin.readline() blocks (even if data is available, but the
line isn't finished).

Resolving that would require something a lot more complicated, such as signals,
threading, or a custom line parser.  This flag is much easier.