tsan: deflake race-order-dependent tests
These tests can have the race occur in any order; as written, they _usually_
happen in the given order, but occasionally don't, causing test flakes.  I found
the eligible tests by running the TSAN tests a bunch of times --- there are
probably more tests to be deflaked in this manner.