tsan: disable tsan during reporting
Also disables logging function entry/exits while TSAN is disabled.

This is not generally an issue, but can cause the tests to flake when many races
are being reported --- the race reports themselves cause a logging overflow, and
subsequent race reports to not find the racing access.

There is no need to log function entry/exit when TSAN is disabled, so this
disables that as well to reduce log pressure.