tsan: fix TSAN_CORE helpers in kthread_queue_*
This was technically broken by the previous commit, but the existing tsan tests
don't trigger the false positive in kthread_queue_insert.  They _are_ triggered
by the block cache tests (and at least once in VFS tests, though that's
non-deterministic).

This fixes the false positive by annotating the helpers correctly, and adds a
multi-threaded thrashing test to the TSAN tests that exercises it.