proc: add test for raw-spinlock impl synchronization
This test verifies the _implementation_ of raw spinlock, even outside TSAN_CORE
mode.  We use some preprocessor trickery to create a TU with the normal
(TSAN-instrumented, including atomics) version of raw_spin_lock() etc, then
write a test that verifies those synchronize.

If we break the raw spinlock impl, then a race is flagged in the raw spinlock
itself rather than the test code --- but that's OK.