user-tests: deflake racy waitpid()/SIGCONT tests
This was a fun one!  When written originally, I don't think this was technically
racy as the kernel was not preemptively multitasked.  Since preemption in
userspace was added, this was technically racy but didn't manifest (that I'm
aware of).

This started failing consistently recently, which I tracked down to a libc (I
think?) change that caused a page boundary between the the fork() and
kill(SIGCONT) calls.  This in turn caused, only on i586, the parent and child to
reverse running order and the test to fail.  The child would process the SIGCONT
and exit entirely before the parent entered waitpid().

The new test forces a certain running order.  It has the benefit of also
improving the runtime of the tests by several hundred ms.  I introduced
artificial yielding in kill() and waitpid() calls which revealed several other
tests with similar brittleness, which I also fixed.