proc: fix wait/exit race
This is a tricky race, and was not surfaced in the (deterministic) existing
tests.  First showed up in sponge-20251108-232324, where the exiting child
segfaulted accessing itself because the parent had already cleaned up the child.
This can be triggered if another process is holding the parent's mutex
(causing the exiting child to yield), and then another child exits causing the
parent to wake up in waitpid().

The core of the fix is simple --- have the child get and lock the parent before
starting to exit, so even if the parent runs in that time it won't see the child
as a ZOMBIE.  Additionally, I have the child hold a reference to itself
explicitly.

The challenging part is dealing with the waitpid() side.  In the previous
version of this, technically any other thread could observe a process in
PROC_INVALID if it happened to get it at the right time.  Ideally, threads can
only observe processes in PROC_ZOMBIE --- and, critically, if they proc_get() a
process, that process is active in the global process table.  Having a reference
to a PROC_ZOMBIE process when another (new) process is already using the same ID
would not be good.

Rather than have other threads deal with this complexity, the thread calling
waitpid() in the parent blocks until it has the only remaining reference to the
zombie child.  This requires a bit of extra signalling logic.  We could consider
setting a bit somewhere that would prevent additional threads from getting
references to the process once finish_waitpid() starts to clean up --- but then
different threads would be observing different states for the same pid, which
might be an issue.

This blocking can cause deadlock if another thread holds a reference to a child
process then tries to lock its parent (which is what used to happen in setpgid()
before commit 4ac34ee65 ("proc: don't lock parent in setpgid()")).