proc: don't lock parent in setpgid()
This was added in 25931830e ("proc: add parent/child hierarchy locking"), but
it's not actually necessary to lock the parent --- we can read the child's
parent field with only the child's mutex locked.

This is the only place where proc_get_and_lock_parent() is called with a process
that isn't the current process.  That is actually unsafe.