proc: proc_get_and_lock_parent() only works on current proc
We no longer need a version of this that works on any process, and in fact that
is dangerous --- a process could transition to PROC_ZOMBIE or PROC_INVALID while
this is running.  Today, that results in a kernel segfault.  An upcoming change
to fix a broader race condition with proc exit makes this deadlock instead of
segfault.

A non-deadlocking version could be written that attempts to stabilize the parent
but if the process is concurrently exiting fails the whole operation.  If and
when we need to acquire both an arbitrary process and its parent simultaneously,
we can write that.