proc: add multi-thread waitpid() test
This exercises the scenario of multiple threads in a parent process calling
waitpid() simultaneously, and racing for who can clean up a child first.  Only
one of the parent threads should succeed.  If waitpid() is updated to, for
example, incorrectly release the parent's lock while it waits for the child
process refcount to go down, this test fails.