proc: add signal tests for process groups with ZOMBIEs
There are already tests for signalling a ZOMBIE, and signalling a process group,
but none that exercise both together.  This also adds explicit tests for
proc_force_signal_group() --- it is already tested in many places by users of it
(e.g. the ld or tty code), but not on its own.