signal tests: fix use-after-free
Amazingly, when the tests are normally run in kernel mode it seems like the
thread block is reallocated at exactly the same address, so the fact that a
stale pointer is used on line 1268 doesn't cause a problem.

When run from bash with 'ktest all', something else causes the memory allocation
patterns to change such that the use-after-free triggers a segfault.