vfs: add unmount fs op, and call on error or unmount
The current code crashes if an ext2 filesystem fails to mount (for example, due
to a missing mount destination directory), or is unmounted.  In both cases,
ext2fs->mounted is never set to false, so there is a CHECK-fail in
ext2_destroy_fs.

This area needs a refactor in general, and I think there should be a symmetrical
mount fs op.