ld: add basic loader binary
This is a stub for the dynamic loader binary, which puts in place several
components (besides the actual dynamic loading logic).  I've figured out the
right incantations to generate a static PIE binary, after a lot of futzing
around before I realized that my custom toolchain doesn't support -static-pie
(it is silently ignored).

It doesn't work right when linking libc statically, but conveniently that isn't
actually needed for the loader --- we don't want a full libc statically linked
in (at this point at least).  Later on if we end up needing much more of libc,
we could consider a two-phase loader (a simple one the kernel loads that is able
to bring in libc dynamically, then run the rest of its loading code), but that
is almost certainly overkill.